changing flow of view
