+++ assets/js/components/display.js
@@ -1,8 +1,6 @@
 const display = (() => {
 
-  const insertSymbol = (target, playerSymbol) =>{
-    console.log(target.id);
-  }
+  const initialTarget = document.querySelector(".card");
 
   const createTable = () => {
     return `<table class="m-3">
@@ -51,9 +49,47 @@ const display = (() => {
             </div>`
   }
 
+  const initialTemplate = () => {
+    return  `<div class="initial-view">
+              <img src="./assets/images/innertictactoe.jpg" class="card-img-top" alt="...">
+              <div class="card-body">
+                <h5 class="card-title">Welcome</h5>
+                <p class="card-text">Click below to start</p>
+                <button type="button" class="btn btn-primary" data-toggle="modal" data-target="#exampleModalCenter">
+                  Start
+                </button>
+                <div class="modal fade" id="exampleModalCenter" tabindex="-1" role="dialog" aria-labelledby="exampleModalCenterTitle" aria-hidden="true">
+                  <div class="modal-dialog modal-dialog-centered" role="document">
+                    <div class="modal-content">
+                      <div class="modal-header">
+                        <h5 class="modal-title" id="exampleModalCenterTitle">Enter Player Names</h5>
+                        <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+                          <span aria-hidden="true">&times;</span>
+                        </button>
+                      </div>
+                      <div class="modal-body">
+                        <form>
+                          <div class="form-group">
+                            <input type="text" class="form-control" id="exampleInputName1" aria-describedby="emailHelp" placeholder="Player X">
+                            <input type="text" class="form-control" id="exampleInputName2" aria-describedby="emailHelp" placeholder="Player Y">
+                          </div>
+                        </form>
+                      </div>
+                      <div class="modal-footer">
+                        <button type="button" class="btn btn-primary" id ="play" data-dismiss="modal">Play</button>
+                      </div>
+                    </div>
+                  </div>
+                </div>
+              </div>
+            </div>`
+  }
+
+
   return {
     createTable,
     createUserInfo,
+    initialTemplate,
   };
 })();
 export default display;
